requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
boto3==1.40.69
botocore==1.40.69
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
fastapi==0.115.2
h11==0.16.0
httptools==0.7.1
idna==3.11
jmespath==1.0.1
numpy==2.2.6
opencv-python==4.12.0.88
passlib==1.7.4
pillow==12.0.0
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.6
PyYAML==6.0.3
requests==2.32.0
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
starlette==0.40.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.23.2
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
